+++ CHANGELOG.md
@@ -9,6 +9,7 @@ We follow [Semantic Versions](https://semver.org/).
 - Update dependencies
 - Extend the `Config` to allow storing more items locally (using YAML)
 - Initialize the `Config` inside the `App` class to handle errors during the `Config` loading
+- Add `Schedule`, `ScheduleStorage` and `ScheduleFileStorage` classes for storing scheduled posts
 
 ## Version 0.1.8